requirements.txt
python-dotenv
pydantic
typing_extensions
langchain_core
langchain_openai
langgraph
pandas
numpy
vanna
langchain_community
langchain_experimental
ipython
chainlit
openpyxl
google-cloud-aiplatform
pdf2image
langchain
langchain-google-vertexai
pypdf
langchain-astradb
cassandra-driver
Pillow
